+++ sdk/src/androidTest/java/ly/count/android/sdk/messaging/PushTests.java
@@ -16,7 +16,6 @@ public class PushTests {
 
     @Before
     public void setUp() {
-        Countly.sharedInstance().setLoggingEnabled(true);
     }
 
     @After